# convertVarIntToBytes: use reusable bytes array not expensive, discarded bytes.Buffer