@odeke-em 2020-12-17
Thank you @erikgrinaker for the review! Could you please help me cut a new release too? I ask because it would be nice for us to get the cosmos-sdk team to use the latest code as well as the changes you made over the weekend fixing identified issues.