@odeke-em 2020-12-20
Gotcha, thanks for the update. If you are able to, I kindly request that you help me cut a point release with this code included, as when I (memory and cpu) profile gaiad, iavl code shows up, and am investigating.